added a few optimizations to make it go faster
I forgot that when you use array lists, each time you add, it doubles the size which means it overestimates the memory needed by a lot. I instead gave it an initial size so that it never calls on resize and this cut down on the time by a lot. I also removed an unnecessary variable in merge which reduced the memory needed by a little bit